ingest/requirements.txt
python-gitlab>=1.10.0
elasticsearch>=7.0.0,<8.0.0
